# config/api-contract.yaml
openapi: 3.1.0
info:
  title: Sparky Gateway API
  version: 0.1.0
  description: |
    Authenticated internal API surface for Sparky (DGX Spark main intelligence
    machine). All external consumers (Maestro, AI Router, MaestroUI, DarkBase
    services, future media/coding agents) call this gateway, not raw runtime
    ports. See PLAN.md §5 for the full contract and §4.1 for the service
    topology.

    Auth: `Authorization: Bearer <SPARKY_API_KEY>`. The key is mirrored from
    MiMi-Secrets (sealed secret `sparky-api-key`) per PLAN §10.

    Endpoints `/health` and `/ready` are unauthenticated for probes. All other routes,
    including `GET /metrics`, require the API key — scrape Prometheus with Bearer auth.

    Request bodies are validated against this schema before enqueue or proxy;
    out-of-policy numeric bounds or unsafe URIs receive HTTP 422 with an error
    envelope (same shape as other gateway errors).
  contact:
    name: Sparky maintainers
    url: https://github.com/oravandres/Sparky
  license:
    name: MIT
servers:
  - url: http://sparky.mimi.local:8080
    description: Internal LAN gateway
security:
  - sparkyApiKey: []
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [text]
      summary: OpenAI-compatible chat completions (proxied to Nemotron)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Completion response (OpenAI-compatible shape)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/reasoning/analyze:
    post:
      tags: [reasoning]
      summary: Single-input deep analysis (PLAN §5.2.1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasoningAnalyzeRequest"
      responses:
        "200":
          description: Structured analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAnalyzeResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/reasoning/compare:
    post:
      tags: [reasoning]
      summary: Side-by-side option comparison (PLAN §5.2.2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasoningCompareRequest"
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningCompareResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
  securitySchemes:
    sparkyApiKey:
      type: http
      scheme: bearer
      description: SPARKY_API_KEY (mirrored from MiMi-Secrets, see PLAN §10)
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unavailable:
      description: Underlying runtime is unavailable (model loading, evicting, etc.)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Request failed schema or policy validation (bounds, URI allowlist, etc.)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: Upstream runtime returned an error or an invalid response envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
...
    BoundedGenerationTokens:
      type: integer
      minimum: 1
      maximum: 16384
      description: >-
        Default production ceiling per generation request before proxy to the always-hot text
        runtime (DoS guard). Larger batches require an explicit api-contract / privileged-tier bump,
        not an unbounded client integer.
    ChatMessage:
      type: object
      required: [role, content]
      additionalProperties: false
      properties:
        role: { type: string, enum: [system, user, assistant, tool] }
        content:
          type: string
          minLength: 1
          maxLength: 1000000
      description: >-
        Parser hard cap per message. The gateway may return HTTP 422 when a message
        exceeds SPARKY_CHAT_MAX_CONTENT_CHARS (default 120000) before proxying.
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      additionalProperties: false
      properties:
        model: { type: string }
        messages:
          type: array
          minItems: 1
          maxItems: 128
          items: { $ref: "#/components/schemas/ChatMessage" }
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_tokens: { $ref: "#/components/schemas/BoundedGenerationTokens" }
        top_p:
          type: number
          minimum: 0
          maximum: 1
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
        stop:
          description: >
            At most four stop sequences of at most 256 characters each (string or array).
          oneOf:
            - type: string
              maxLength: 256
            - type: array
              maxItems: 4
              items:
                type: string
                maxLength: 256
        user:
          type: string
          maxLength: 256
        stream:
          type: boolean
          enum: [false]
          description: >
            SSE streaming is not implemented yet; only false may be sent if the
            field is present. Omit this field for non-streaming OpenAI clients.
      description: >
        Whitelisted OpenAI-compatible non-streaming chat fields only; additional
        properties are rejected with HTTP 422. The gateway proxies this payload
        to the Nemotron text runtime. Clients must not send stream=true until the
        gateway implements SSE `text/event-stream` chunks.
    ReasoningAnalyzeRequest:
      type: object
      required: [task]
      properties:
        task: { type: string }
        context: { type: string }
        criteria:
          type: array
          items: { type: string }
        output_style: { type: string, enum: [structured, prose, bulleted] }
        max_tokens:
          type: integer
          minimum: 1
          maximum: 16384
          default: 2048
    ReasoningAnalyzeResponse:
      type: object
      required: [summary, key_points, risks, assumptions, recommendation, confidence]
      properties:
        summary: { type: string }
        key_points: { type: array, items: { type: string } }
        risks: { type: array, items: { type: string } }
        assumptions: { type: array, items: { type: string } }
        recommendation: { type: string }
        confidence: { type: string, enum: [high, medium, low] }
    ReasoningCompareRequest:
      type: object
      required: [question, options, criteria]
      properties:
        question: { type: string }
        options:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id: { type: string }
              name: { type: string }
              description: { type: string }
        criteria:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id: { type: string }
              name: { type: string }
              weight: { type: number, default: 1.0 }
        constraints:
          type: array
          items: { type: string }
    ReasoningCompareResponse:
      type: object
      required: [scores, totals, recommendation, confidence]
      properties:
        scores:
          type: array
          items:
            type: object
            required: [option_id, criterion_id, score, rationale]
            properties:
              option_id: { type: string }
              criterion_id: { type: string }
              score: { type: number }
              rationale: { type: string }
        totals:
          type: array
          items:
            type: object
            required: [option_id, weighted_total]
            properties:
              option_id: { type: string }
              weighted_total: { type: number }
        recommendation:
          type: object
          required: [option_id, reasoning]
          properties:
            option_id: { type: string }
            reasoning: { type: string }
            caveats:
              type: array
              items: { type: string }
        confidence: { type: string, enum: [high, medium, low] }
...